requirements.txt
langchain-openai
langchain
python-dotenv
pandas
openai
lancedb
ipykernel